Split resolver declaration definition
